Use setuptools entry_points to register pytest plugins
So that they work consistently for editable and site-packages installations
